Move InspectorTarget to modules
