+++ aixplain/factories/team_agent_factory/__init__.py
@@ -23,25 +23,15 @@ Description:
 
 import json
 import logging
-from enum import Enum
 from typing import Dict, List, Optional, Text, Union
 from urllib.parse import urljoin
 
 from aixplain.enums.supplier import Supplier
 from aixplain.modules.agent import Agent
-from aixplain.modules.team_agent import TeamAgent
+from aixplain.modules.team_agent import TeamAgent, InspectorTarget
 from aixplain.utils import config
 from aixplain.factories.team_agent_factory.utils import build_team_agent
 from aixplain.utils.file_utils import _request_with_retry
-
-
-class InspectorTarget(str, Enum):
-    # TODO: INPUT
-    STEPS = "steps"
-    OUTPUT = "output"
-
-    def __str__(self):
-        return self._value_
 
 
 class TeamAgentFactory:
